backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
flask==3.0.3
flask-restx==1.3.0
flask-socketio==5.3.6
flask-mail==0.10.0
flask-jwt-extended==4.6.0
flask-limiter==3.7.0
flask-cors==4.0.1
flask-bcrypt==1.0.1
flask-sqlalchemy==3.1.1
gunicorn==22.0.0

# ── Async / queue ─────────────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.4
eventlet==0.36.1

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30

# ── ML / CV ───────────────────────────────────────────────────────────────────
torch==2.3.0
torchvision==0.18.0
timm==1.0.3
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.13.0

# ── XAI ───────────────────────────────────────────────────────────────────────
lime==0.2.0.1
shap==0.45.1

# ── Reports ───────────────────────────────────────────────────────────────────
reportlab==4.2.0
pandas==2.2.2
matplotlib==3.8.4

# ── Integrations ──────────────────────────────────────────────────────────────
python-telegram-bot==21.3
requests==2.32.2
openai==1.30.1

# ── Utils ─────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pyjwt==2.8.0
bcrypt==4.1.3
marshmallow==3.21.2
